feat(cli/settings): --goto/--help/--version, files.associations, save mkdir -p, goto line validation
- cli.v: new parse_cli_args with --help / --version / -g FILE:LINE[:COL] /
  literal '-' stdin; main() routes --help/--version to print-and-exit so
  they never touch the terminal (matches Rust main.rs).
- main.v: add_document accepts missing paths (creates an empty buffer
  pointed at the path, sets has_file_id=false) and normalises via
  os.abs_path; save_active mirrors Rust's open_for_writing with mkdir -p
  for nested targets. Ctrl+G prompt uses parse_prompt_goto + restart on
  parse failure, supports 'LINE' and 'LINE:COLUMN', clamps column at line
  end via cursor_move_to_logical.
- filepicker.v: picker_do_save also mkdir -p on the target directory.
- settings.v: load_settings reads ~/.config/edit/settings.json and
  honours files.associations; bare globs auto-prefix **/; unknown
  language ids / bad JSON go to the editor's error log instead of
  aborting.
- highlighter.v / language_picker.v: language_for_path consults user
  associations first, then the built-in ext->language table.
- menubar.v: About dialog uses injected edit_version; build.sh reads
  the version from v.mod and passes -d edit_version=... to v.
- sys.v: macos termios flag mask uses int (darwin TcFlag is usize, V
  reads it as int).
- tests: cli_test.v covers parse_cli_args / parse_filename_goto /
  parse_prompt_goto / goto_line_index; settings_test.v covers glob
  normalisation and JSON load paths; main_test.v covers Ctrl+G goto
  + save mkdir-p + missing-path add_document; goto_file_test.v adds
  add_goto_test_document helper and now calls goto_file_compute_filtered
  so the list_h-driven scroll tests see the right index set.
- docs: TODO.md ticks the new prompt/CLI/search items; PLAN.md is a new
  compatibility roadmap scoped to the V port (P0 CLI+goto+settings, P1
  regex/Unicode, P2 focus+perf).

+++ edit/sys.v
@@ -172,12 +172,13 @@ pub fn switch_modes() ! {
 
 	$if macos {
 		// V 编译器在 import os 时对 `u64(C.A | C.B)` 的组合表达式有 bug，
-		// 逐项单独转换再或运算可以绕开。
-		iflag_mask := u64(C.IGNBRK) | u64(C.BRKINT) | u64(C.PARMRK) | u64(C.INPCK) | u64(C.ISTRIP) | u64(C.INLCR) | u64(C.IGNCR) | u64(C.ICRNL) | u64(C.IXON)
-		oflag_mask := u64(C.OPOST)
-		cflag_mask := u64(C.CSIZE) | u64(C.PARENB)
-		lflag_mask := u64(C.ISIG) | u64(C.ICANON) | u64(C.ECHO) | u64(C.ECHONL) | u64(C.IEXTEN)
-		cs8 := u64(C.CS8)
+		// 逐项单独转换再或运算可以绕开。`struct C.termios` 的 TcFlag
+		// 在 darwin 上是 usize，在 v 0.5.2 下编译器读为 int，所以用 int。
+		iflag_mask := int(C.IGNBRK) | int(C.BRKINT) | int(C.PARMRK) | int(C.INPCK) | int(C.ISTRIP) | int(C.INLCR) | int(C.IGNCR) | int(C.ICRNL) | int(C.IXON)
+		oflag_mask := int(C.OPOST)
+		cflag_mask := int(C.CSIZE) | int(C.PARENB)
+		lflag_mask := int(C.ISIG) | int(C.ICANON) | int(C.ECHO) | int(C.ECHONL) | int(C.IEXTEN)
+		cs8 := int(C.CS8)
 
 		termios.c_iflag &= ~iflag_mask
 		// Disable output processing.
